Re-export hyper::Uri as shiplift::Uri
This patch removes the explicit dependency on the http crate and
instead accesses the re-exported version of hyper. This should make
the update process slightly easier because those versions would
need to be kept in sync manually.

We also re-export hyper::Uri as shiplift::Uri because it is part
of the public API of shiplift::Docker. This allows users to access
the Uri type without having to seperately depend on http or hyper.